Add files via upload
This model provides our proposed Model-Based Systems Engineering (MBSE) method for SoAS architecture and integration implemented on a hypothetical search-and-rescue case study. The proposed method falls under the System Engineering for AI (SE4AI) umbrella where SE principles are tailored to accommodate challenges posed by the integration of autonomy. The model guides the identification and creation of multiple SoAS architectures in the Unified Architecture Framework (UAF) that must be analyzed for suitability and emergent behavior.